requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
geopy==2.4.1
flask-talisman==1.1.0
flask-wtf==1.2.1
flask-limiter==3.7.0
waitress==3.0.0
pytest==8.2.1
pytest-flask==1.3.0
email-validator==2.1.1
